SmsQueueSender: rework and improve.
the SMS Queue Sender reads the Osmo internal SMS database and
looks for pending SMS for user who are roaming elsewhere.

Update to use SMPP to forward the message, which should be much
more reliable.

This commit also introduces much improved and verbose (if desired)
and also colour logging and cmd line options.

+ General code improvement (pylint)