Add the RCCN ESME
The RCCN ESME is using python-smpplib to connect to the OSMO
SMSC, be it in the osmo-nitb, osmo-msc or whatever and it will
process deliver_sm by running the same kind of auth and routing
checks that the SMSC->Kannel->RAPI->Kannel->SMSC routine used to
do.

We reject right away messages to TON INTL and will not accept them
from the phone. The user will see an error.

Note that currently, if we cannot deliver a message because the
remote destination is unavailable, we will not accept it from the
phone. This will be fixed when full queueing is implemented, but
for now it's probably better than accepting the messages that we
probably cannot then deliver.