Implement possibility to activate Jitter Buffer in FS
If there is a lot of jitter on the vpn / voip facing IP connection, it
may make sense to use FreeSwitch's jitter buffer on these call legs.
It appears from some internet posts and some testing that these export
commands are the (only?) way to do it.

The use of nolocal: in the outbound call should ensure that we do
not also have a jitter buffer on the BTS facing leg.
For the inbound call, I'm not sure what FS is doing - in fact I'm
not 100% sure about any of this, but initial reports are that perceived
call quality has improved in the case of a high jitter connection to VOIP
upstream.