Dialplan: Add function to call Free (unbilled) numbers.
We can now define a list of numbers that can be called
without billing the user. Even users that are not
authorised can call these numbers.

The list is in config_values.py named free_numbers

The list should be 10 digit numbers (for Mexico) the
prefix will be added before sending to upstream VOIP
provider.