rccn/rapi: Reformat log generation statements
Modified due to upstream changes by <keith@rhizomatica.org>

This commit refactors the log generation statements in the rapi server
to use the logging module's lazy string evaluation (improving
performance when log levels are disabled) and built-in exception
logging. This commit additionally makes the log statement formatting
consistent.

This commit does *not* change the content of the log statments in any
way.