prefix '+' to international caller IDs
If calls coming in from VoIP providers have international caller id numbers that
start with the country code, then we need to prefix these with a '+' before
forwarding to the GSM side.

Otherwise, the numbers will not display correctly on phones, they may not match
names from the users phone book, and dialing via missed call logs and such
is broken.