rccn/subscriber: Update edit implementation to avoid introspection
The edit implementation previously coupled the database column names
to all function local variable names. It would be easy for a new
developer to introduce a bug simply by moving the code, since the code
is implicitly dependent on the local environment.

This commit explicitly sets the mapping of varaibles to column names,
making it more explicit which columns are modified. This allows for
the column names to change in the future without requiring a change to
the high-level internal api as well.