rccn: Refactor hlr and msc access to separate modules
To isolate the code that is actually interacting with the osmocom
network components, this commit moves osmocom interface code into a
separate adaptor class from the main subscriber. This commit intends
to not change behavior, and otherwise replicate all existing behavior
in the OsmoNitb class.

This commit lays the foundation for components to maintain a stable
internal api while the access to the component databases and
interfaces will change with changes to the network structure.

Using some knowledge from the future, it does *not* abstract away
which 3GPP component provides which information, even if in the nitb
implementation multiple components come from the same process. This
will make it easier to understand where information is coming from if
the components are ever distributed and avoids adding another layer of
indirection. If desired though a wrapper class could be created which
hides which components provide which information to the subscriber.